minor refactoring and suggestions left
